+++ changelog.md
@@ -1,6 +1,8 @@
 # Changelog
 
 
+* [Added] Download not completed data dir
+* [Added] Error log information
 * [Added] Common excel file download
 * [Added] Tags
 * [Added] Refresh button